requirements_txt.txt
fastapi
uvicorn
pydantic
python-dotenv
gradio
langchain
langchain-core
langchain-groq
langchain-ollama
langchain-openai
langchain-community
langchain-huggingface
chromadb
pypdf
tiktoken
requests
sentence-transformers
torch
transformers